Projeto-em-Laravel---Coffee-Website novo

+++ database/migrations/2022_04_24_192428_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('price');
+            $table->decimal('price', 9, 3);
             $table->integer('qty');
             $table->text('description');
             $table->timestamps();
